requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
tqdm
pandas>=2.2
pillow
pytesseract
opencv-python
streamlit>=1.50
scikit-learn>=1.5
fuzzywuzzy
python-Levenshtein
rapidfuzz>=3.0
unidecode>=1.3
numpy>=1.26
altair>=5
sklearn
